Move test classes to a unified package.
All test classes were moved to the `dev.tori.wraith.test` package to standardize organization and improve maintainability. The copyright headers were also updated to reflect the extended range (2024-2025).
