+++ src/test/java/dev/tori/wraith/test/ListenerTargetingTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 7orivorian.
+ * Copyright (c) 2024-2025 7orivorian.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,7 +19,7 @@
  * THE SOFTWARE.
  */
 
-package dev.tori.wraith.eventbus.targeting;
+package dev.tori.wraith.test;
 
 import dev.tori.wraith.bus.EventBus;
 import dev.tori.wraith.event.Target;
