+++ src/test/java/dev/tori/wraith/test/TaskExecutorTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2024 7orivorian.
+ * Copyright (c) 2021-2025 7orivorian.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,7 +19,7 @@
  * THE SOFTWARE.
  */
 
-package dev.tori.wraith.taskexecutor;
+package dev.tori.wraith.test;
 
 import dev.tori.wraith.task.ScheduledTask;
 import dev.tori.wraith.task.TaskExecutor;
